Add support for request batching to UDP Driver
The UDP protocol supports a maximum payload size of ~64kbs. To ensure that we don't lose metrics when attempting to send a large amount, we can send the metrics in batches as long as we don't split up individual lines. Implemented this change into the UDP Driver and added test coverage.